refactor: rename `ESBuildMinifyPlugin` to `EsbuildPlugin` (#307)
BREAKING CHANGE: `ESBuildMinifyPlugin` renamed to `EsbuildPlugin`; `MinifyPluginOptions` renamed to `EsbuildPluginOptions`
